merge

+++ build/13-uninstall_webapps.sh
@@ -5,5 +5,6 @@ for fname in $(ls ../../gs-webbrowser/Applications|grep .app$); do
   rm -rf ${GNUSTEP_LOCAL_WEB_APPS}/${fname}
 done
 
-rmdir ${GNUSTEP_LOCAL_WEB_APPS}
-
+if [[ -z $(ls ${GNUSTEP_LOCAL_WEB_APPS}) ]]; then
+  rmdir ${GNUSTEP_LOCAL_WEB_APPS}
+fi
